voice_service/pyproject.toml
[tool.poetry]
name = "voice_service"
version = "0.1.0"
description = "Service for cloning and cataloguing voices"
authors = ["Manus AI"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.100"
uvicorn = {extras = ["standard"], version = "^0.22"}
pydantic = "^2.0"
httpx = "^0.24"
psycopg2-binary = "^2.9.5" # For Postgres
sqlalchemy = "^2.0"
elevenlabs = "*"  # Added ElevenLabs SDK
python-multipart = "0.0.20"
python-dotenv = "^1.0.0"  # For loading environment variables
[tool.poetry.dev-dependencies]
pytest = "^7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
